feat: 북마크 및 알림 아티클 삭제 API 경로 수정 및 쿼리 수정
