+++ apps/client/src/pages/remind/Remind.tsx
@@ -5,11 +5,14 @@ import OptionsMenuPortal from '@shared/components/sidebar/OptionsMenuPortal';
 import { useAnchoredMenu } from '@shared/hooks/useAnchoredMenu';
 import { belowOf } from '@shared/utils/anchorPosition';
 import { REMIND_MOCK_DATA } from './constants';
-import { useDeleteRemindArticle, useGetRemindArticles } from './apis/queries';
+import { useGetRemindArticles } from './apis/queries';
 import { formatLocalDateTime } from '@shared/utils/formatDateTime';
 import NoReadArticles from '@pages/remind/components/noReadArticles/NoReadArticles';
 import NoUnreadArticles from '@pages/remind/components/noUnreadArticles/NoUnreadArticles';
-import { usePutArticleReadStatus } from '@shared/apis/queries';
+import {
+  useDeleteRemindArticle,
+  usePutArticleReadStatus,
+} from '@shared/apis/queries';
 import { useQueryClient } from '@tanstack/react-query';
 
 const Remind = () => {
