+++ apps/client/src/pages/remind/apis/axios.ts
@@ -11,8 +11,3 @@ export const getRemindArticles = async (
   );
   return data.data;
 };
-
-export const deleteRemindArticle = async (id: number) => {
-  const response = await apiRequest.delete(`/api/v1/articles/${id}`);
-  return response;
-};
