+++ apps/client/src/shared/apis/axios.ts
@@ -29,9 +29,9 @@ export const getAcorns = async () => {
 };
 
 export interface postSignUpRequest {
-  email: string,
-  remindDefault: string,
-  fcmToken: string | null
+  email: string;
+  remindDefault: string;
+  fcmToken: string | null;
 }
 
 export const postSignUp = async (responsedata: postSignUpRequest) => {
@@ -48,5 +48,10 @@ export const putArticleReadStatus = async (articleId: number) => {
 
 export const deleteCategory = async (id: number) => {
   const response = await apiRequest.delete(`/api/v1/categories/${id}`);
+  return response;
+};
+
+export const deleteRemindArticle = async (id: number) => {
+  const response = await apiRequest.delete(`/api/v1/articles/${id}`);
   return response;
 };
