+++ apps/client/src/shared/apis/queries.ts
@@ -13,6 +13,7 @@ import {
   putCategory,
   getAcorns,
   putArticleReadStatus,
+  deleteRemindArticle,
 } from '@shared/apis/axios';
 import { AxiosError } from 'axios';
 import {
@@ -81,5 +82,11 @@ export const usePutArticleReadStatus = (): UseMutationResult<
 > => {
   return useMutation({
     mutationFn: (articleId: number) => putArticleReadStatus(articleId),
+  });
+};
+
+export const useDeleteRemindArticle = () => {
+  return useMutation({
+    mutationFn: (id: number) => deleteRemindArticle(id),
   });
 };
